Experimental things
* Multitouch handling, allows one of these actions below to work with
moving camera at the same time.
- Switching hotbar slot; might not work properly
- Destroying block (left mouse button) with another finger
long-pressed (without pressing the PRI button)
- Right mouse trigger with another finger tapped onto the screen
(without pressing the SEC button); not working properly yet
* Sub-pixel cursor support; not yet checked

+++ JavaApp/src/main/java/net/kdt/pojavlaunch/uikit/UIKit.java
@@ -104,7 +104,7 @@ public class UIKit {
         PLaunchApp.launchMinecraft();
     }
     
-    public static void callback_SurfaceViewController_onTouch(int event, int x, int y) {
+    public static void callback_SurfaceViewController_onTouch(int event, float x, float y) {
         switch (event) {
             case ACTION_DOWN:
             case ACTION_UP:
